requirements.txt
altair>=5.5.0
arrow>=1.3.0
attrs>=25.1.0
blinker>=1.9.0
branca>=0.8.1
cachetools>=5.5.1
certifi>=2024.12.14
charset-normalizer>=3.4.1
click>=8.1.8
contourpy>=1.3.1
cycler>=0.12.1
elastic-transport>=8.17.0
elasticsearch>=8.17.1
folium>=0.19.4
fonttools>=4.56.0
geographiclib>=2.0
geopy>=2.4.1
gitdb>=4.0.12
GitPython>=3.1.44
idna>=3.10
Jinja2>=3.1.5
joblib>=1.4.2
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
kiwisolver>=1.4.8
kneed>=0.8.5
loguru>=0.7.3
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
matplotlib>=3.10.0
mdurl>=0.1.2
mycolorpy>=1.5.1
narwhals>=1.27.1
numpy>=2.2.2
packaging>=24.2
pandas>=2.2.3
pillow>=11.1.0
polyline>=2.0.2
proj>=0.2.0
protobuf>=5.29.3
PuLP>=2.9.0
pyarrow>=19.0.1
pydeck>=0.9.1
Pygments>=2.19.1
pyparsing>=3.2.1
python-dateutil>=2.9.0.post0
pytz>=2024.2
PyYAML>=6.0.1
referencing>=0.36.2
requests>=2.32.3
rich>=13.9.4
rpds-py>=0.22.3
Rtree>=1.3.0
scikit-learn>=1.6.1
scipy>=1.15.2
six>=1.17.0
smmap>=5.0.2
streamlit>=1.42.1
tenacity>=9.0.0
threadpoolctl>=3.5.0
toml>=0.10.2
tornado>=6.4.2
types-python-dateutil>=2.9.0.20241206
typing_extensions>=4.12.2
tzdata>=2025.1
urllib3>=2.3.0
xyzservices>=2025.1.0
networkx>=3.5